Initial migration of tests to typescript
